Add namespaces
This is the annoying (but necessary) code that ensures that these
methods will not accidentally collide with other methods elsewhere
in the app that have the same name.

Note the code doesn't change when it gets namespaced, it just gets
surrounded in more code.

So the un-namespaced method:

    def foo
    end

Becomes:

    module Namespace
      def foo
      end
    end

This is nice because it means you don't have to think about namespaces
up-front when sketching something out.

The `module_function` macro is used to indicate that the method is
actually a pure function and so doesn't need to be "mixed in" to some
class before it can be used.

Module functions can be called directly on the module, e.g.:

    Namespace.foo()